requirements.txt
asgiref==3.6.0
Django==4.1.5
django-environ==0.9.0
django-nose==1.4.7
factory-boy==3.2.1
Faker==16.6.0
nose==1.3.7
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
